requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
boto3==1.39.4
python-multipart==0.0.6
httpx==0.25.2 